+++ src/writer.rs
@@ -9,9 +9,12 @@ pub struct GpsWriter {}
 impl GpsWriter {
 
     /// Writes a PMTK command or query to the UART.
-    pub async fn send<UART: Write + ErrorType>(&mut self, uart: &mut UART, command: impl CmdQ) -> Result<(), GpsError<UART::Error>> {
+    ///
+    /// * `uart` - the UART driver to write to.
+    /// * `cmd_q` - the PMTK command or query to write to the UART.
+    pub async fn send<UART: Write + ErrorType>(&mut self, uart: &mut UART, cmd_q: impl CmdQ) -> Result<(), GpsError<UART::Error>> {
         #[cfg(feature = "defmt")]
-        debug!("GpsWriter.send: {}", command.serialize()?.as_bytes());
-        Ok(uart.write_all(command.serialize()?.as_bytes()).await.map_err(GpsError::Uart)?)
+        debug!("GpsWriter.send: {}", cmd_q.serialize()?.as_bytes());
+        Ok(uart.write_all(cmd_q.serialize()?.as_bytes()).await.map_err(GpsError::Uart)?)
     }
 }